+++ PaddleNLP/paddlenlp/metrics/__init__.py
@@ -15,3 +15,4 @@
 from .perplexity import Perplexity
 from .chunk import ChunkEvaluator
 from .bleu import BLEU
+from .glue import AccuracyAndF1, Mcc, PearsonAndSpearman
